More reliable card display
Moved card wrapper to render loop.
